requirements.txt
# HTTP client
httpx==0.27.0

# Vector database
qdrant-client==1.9.1

# Embeddings
sentence-transformers==3.0.1
torch==2.3.1

# Chunking / RAG framework
llama-index-core==0.10.43
tiktoken==0.7.0

# Document parsing & cleaning
docling==2.4.0

# Settings & validation
pydantic-settings==2.2.1

# CLI
typer==0.12.3
rich==13.7.1

# Dev utilities
python-dotenv==1.0.1
